Issues #127 feat: SettingFragment에 설정 관련 Fragment가 있는 구조로 변경
같잖은 event와 sharedflow 갖다버리고 리스너로 광명 찾았습니다.
